refactor cons usage in lispm.c
Replace the antipattern of triple assignment with a more lean option.
The change is breaking and lrt0.c is in a state of disarray, I would
however like to commit the already done changes.

I am optimistic about the impact on the binary size, I expect the
optimizer to remove all the syntactic fluff.

Several questions raised as I was passing through the code, I plan to
address them before merging this branch into main.