another optimization pass: 80 bytes less
Large chunk of the savings is thanks to changing `lispm_st_obj_alloc()`
to merely move the stack pointer, and off-loading the object
initialization to the caller.

Removing redundant checks in `evapply()` saved some more.

The utility `parse_pair()` actually slightly pessimized the output with
my local gcc, but I left it in nevertheless as I find that it better
organizes the parser code.

My favorite is the finding in `cons_reverse_inplace()`, where I realized
that `M.stack + ...` evaluation is unnecessary and yields `cons[1]`.