extract semantic analysis in its own phase
Essentially paying the price of the past mistakes. Mixing parsing and
semantic analysis was okay-ish previously, but it stays in the way of
implementing `letrec`, because the way I envision the latter is by
transforming `(letrec ((a abody) (b bbody) body)` into
`((lambda (a b) body) (abody bbody))` but in a way that binds `a` and
`b` into the definitions of `abody` and `bbody`, and for that I should
scan all the names before going into the semantic analysis of the
bodies.