step 9: introduce '(apply)' special form
This makes the program after semantic stage look very uniform:
it is either an atom, or a pair `(form args)`, where `form` is a special
form, and args are its parameters.

In case of the `(apply)` form specifically the first argument can either
be a builtin function, or a lambda, it can never be a special form
itself, which simplifies the logic of evapply drastically.