step 10: patch a bug in letrec
Previously `(let ((c 3)) (letrec ((c c)) c))` returned 3, because
`(c c)` bound the value to the one in the containing frame. This is
wrong behavior, because expressions on the right hand side of
assignments in `letrec` should already refer to the objects defined in
`letrec`. In this particular example, the evaluation should cause an
error (either infinite recursion, or undefined symbol).

This can be achieved by unbinding all the `letrec` names before passing
them into the implementation lambda.