an extremely bare to the bone REPL
The REPL works, but keeps the state split between the LISP world and the
native environment. For instance, the result of `#io:readline` is not a
genuine value, and the result of `#eval ln` depends only on the internal
state obtained during `io_readline`. I miss some capabilities for proper
implementation, mainly a natural way to bind a VM value to a
behind-the-scene state.